feat: add working findPassword
Main bug was that I was passing in the wrong "hash" variable to the
findNode function. Instead of passing the original hash, was passing in
the reduced hash (which meant I was always passing the final hash in the
chain).

Main refactor was decoupling findChains from findNode. Now, findChains
returns a list of possible chains, which then get passed to findNode.

Some other minor changes like having to add a mySafeHead function.